requirements.txt
altgraph==0.16.1
apparmor==2.13.1
asn1crypto==0.24.0
astroid==2.2.5
Brlapi==0.6.6
certifi==2017.11.5
chardet==3.0.4
cryptography==2.3
cupshelpers==1.0
cycler==0.10.0
future==0.17.1
glinux-homedir-helper==1
glinux-rebootd==0.1
goobuntu-config-tools==0.1
googletrans==2.4.0
idna==2.6
IPy==0.83
isort==4.3.15
keyring==10.6.0
keyrings.alt==2.2
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
LibAppArmor==2.13.1
louis==3.7.0
macholib==1.11
matplotlib==3.0.3
mccabe==0.6.1
monotonic==1.0
numpy==1.13.1
obno==32
onboard==1.4.1
pandas==0.24.2
pefile==2018.8.8
Pillow==5.4.1
pkg-resources==0.0.0
psutil==5.4.2
pycairo==1.15.4
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
pygame==1.9.4
pygobject==3.26.1
pyinotify==0.9.6
PyInstaller==3.4
pylint==2.3.1
pyOpenSSL==17.5.0
pyparsing==2.3.1
pyperclip==1.7.0
pysmbc==1.0.15.6
python-apt==1.8.3
python-dateutil==2.8.0
python-debian==0.1.31
python-xapp==1.0.0
python-xlib==0.20
pytz==2018.9
pyxdg==0.25
PyYAML==3.12
requests==2.20.0
scour==0.36
SecretStorage==2.3.1
setproctitle==1.1.10
six==1.12.0
twstock==1.2.1
typed-ast==1.3.1
ufw==0.35
urllib3==1.24
virtualenv==15.1.0
wrapt==1.11.1
youtube-dl==2017.12.31
